+++ src/main/java/org/jsoup/select/Collector.java
@@ -3,8 +3,8 @@ package org.jsoup.select;
 import org.jsoup.nodes.Element;
 import org.jspecify.annotations.Nullable;
 
-import java.util.Optional;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * Collects a list of elements that match the supplied criteria.
@@ -16,17 +16,26 @@ public class Collector {
     private Collector() {}
 
     /**
-     Build a list of elements, by visiting root and every descendant of root, and testing it against the evaluator.
+     Build a list of elements, by visiting the root and every descendant of root, and testing it against the Evaluator.
      @param eval Evaluator to test elements against
      @param root root of tree to descend
      @return list of matches; empty if none
      */
-    public static Elements collect (Evaluator eval, Element root) {
-        eval.reset();
+    public static Elements collect(Evaluator eval, Element root) {
+        return stream(eval, root).collect(Collectors.toCollection(Elements::new));
+    }
 
-        return root.stream()
-            .filter(eval.asPredicate(root))
-            .collect(Collectors.toCollection(Elements::new));
+    /**
+     Obtain a Stream of elements by visiting the root and every descendant of root and testing it against the evaluator.
+
+     @param evaluator Evaluator to test elements against
+     @param root root of tree to descend
+     @return A {@link Stream} of matches
+     @since 1.19.1
+     */
+    public static Stream<Element> stream(Evaluator evaluator, Element root) {
+        evaluator.reset();
+        return root.stream().filter(evaluator.asPredicate(root));
     }
 
     /**
@@ -37,9 +46,6 @@ public class Collector {
      @return the first match; {@code null} if none
      */
     public static @Nullable Element findFirst(Evaluator eval, Element root) {
-        eval.reset();
-
-        Optional<Element> first = root.stream().filter(eval.asPredicate(root)).findFirst();
-        return first.orElse(null);
+        return stream(eval, root).findFirst().orElse(null);
     }
 }